+++ blocksatcli/api/msg.py
@@ -3,6 +3,7 @@ import sys, os, logging, time, struct, zlib
 from math import ceil, floor
 import zfec
 from . import pkt
+from .fec import Fec
 
 
 logger            = logging.getLogger(__name__)
@@ -13,13 +14,6 @@ logger            = logging.getLogger(__name__)
 MSG_HEADER_FORMAT = '<255sxI'
 MSG_HEADER_LEN    = 255 + 1 + 4
 data_formats      = ["original", "encapsulated", "encrypted", "fec_encoded"]
-# FEC chunk header:
-# Octet 0      : Chunk id
-# Octets 1 - 4 : Message length
-FEC_HEADER_FORMAT = '!BI'
-FEC_HEADER_LEN    = 5
-# Each FEC chunk goes in a single Blocksat Packet along with metadata
-FEC_CHUNK_SIZE    = pkt.MAX_PAYLOAD - FEC_HEADER_LEN
 
 
 class ApiMsg:
@@ -277,55 +271,8 @@ class ApiMsg:
             overhead : Percentage of FEC chunks to add as overhead (rounded up).
 
         """
-        # Check if the data fits into 256 chunks with the chosen overhead
-        data              = self.get_data()
-        n_chunks          = ceil(len(data) / FEC_CHUNK_SIZE)
-        n_overhead_chunks = ceil(overhead * n_chunks)
-        n_fec_chunks      = n_chunks + n_overhead_chunks
-        logger.debug("FEC Encoder: Original Chunks: {} / "
-                     "Overhead Chunks: {}".format(n_chunks, n_overhead_chunks))
-        if (n_fec_chunks >= 256):
-            max_size = floor(256/(1 + overhead)) * FEC_CHUNK_SIZE
-            raise ValueError("Message size is too large - max size with {} "
-                             "FEC overhead is {:.2f} kB".format(
-                                 overhead, max_size/(2**10)))
-
-        # Split the original data (from the highest container level) into chunks
-        chunks   = []
-        for i_frag in range(n_chunks):
-            # Byte range of the next chunk:
-            s_byte = i_frag * FEC_CHUNK_SIZE # starting byte
-            e_byte = (i_frag + 1) * FEC_CHUNK_SIZE # ending byte
-            chunk  = data[s_byte:e_byte]
-
-            # The last chunk may need zero-padding
-            if (i_frag + 1 == n_chunks and len(chunk) < FEC_CHUNK_SIZE):
-                chunk += bytes(FEC_CHUNK_SIZE - len(chunk))
-
-            assert(len(chunk) == FEC_CHUNK_SIZE)
-            chunks.append(chunk)
-
-        # Generate the corresponding FEC chunks
-        encoder    = zfec.Encoder(n_chunks, n_fec_chunks)
-        fec_chunks = encoder.encode(chunks)
-
-        # Form the payload to be transmitted on each Blocksat Packet. Each
-        # payload contains the FEC chunk and metadata. The metadata includes the
-        # chunk id and the original message length, which the receiver needs.
-        payloads = []
-        for i_chunk, chunk in enumerate(fec_chunks):
-            metadata = struct.pack(FEC_HEADER_FORMAT, i_chunk, len(data))
-            payloads.append(metadata + chunk)
-
-        # Concatenate all payloads to form a single "API message". Later on, the
-        # satellite transmitter splits this message into packets. The FEC chunks
-        # are sized so that each packet derived in this process will carry
-        # precisely a single FEC chunk.
-        message = b""
-        for payload in payloads:
-            message += payload
-
-        self.data['fec_encoded'] = message
+        fec = Fec(overhead)
+        self.data['fec_encoded'] = fec.encode(self.get_data())
 
     def fec_decode(self):
         """Forward error correction (FEC) decoding
@@ -339,66 +286,14 @@ class ApiMsg:
             "is_fec_decodable()" method before calling it.
 
         """
-        encoded_data = self.data['fec_encoded']
-        assert(encoded_data is not None)
+        assert(self.data['fec_encoded'] is not None)
 
-        # The encoded data must contain an integer number of max-sized payloads,
-        # although it may not contain the full FEC-encoded object (given that
-        # some parts of it may have been lost)
-        assert(len(encoded_data) % pkt.MAX_PAYLOAD == 0)
+        fec = Fec()
+        decoded_data = fec.decode(self.data['fec_encoded'])
 
-        # Process each payload and create a map of FEC chunks
-        n_payloads = len(encoded_data) // pkt.MAX_PAYLOAD
-        chunk_map  = {}
-        msg_len    = None
-        for i_payload in range(n_payloads):
-            # Byte range of the next payload:
-            s_byte  = i_payload * pkt.MAX_PAYLOAD # starting byte
-            e_byte  = (i_payload + 1) * pkt.MAX_PAYLOAD # ending byte
-            payload = encoded_data[s_byte:e_byte]
-
-            # Unpack the metadata from the FEC header
-            metadata = struct.unpack(FEC_HEADER_FORMAT,
-                                     payload[:FEC_HEADER_LEN])
-            chunk_id = metadata[0]
-
-            # All payloads should bring the same message length
-            if (msg_len is not None):
-                assert(metadata[1] == msg_len)
-            else:
-                msg_len = metadata[1]
-
-            # Save the FEC chunk
-            chunk_map[chunk_id] = payload[FEC_HEADER_LEN:]
-
-        logger.debug("FEC Decoder: Processed chunks: {}".format(n_payloads))
-
-        # Number of FEC chunks required to decode the original message:
-        n_chunks  = ceil(msg_len / FEC_CHUNK_SIZE)
-
-        if (n_chunks > len(chunk_map)):
-            raise RuntimeError("Insufficient number of FEC chunks")
-
-        # FEC decoding
-        decoder        = zfec.Decoder(n_chunks, 256)
-        # NOTE: The hard-coded "256" represents the maximum number of chunks
-        # that can be generated. The receiver does not know how many chunks the
-        # sender really generated. Nevertheless, it does not need to know, as
-        # long as it receives at least n_chunks (any combination of
-        # n_chunks). The explanation for this requirement is that the FEC scheme
-        # is a maximum distance separable (MDS) erasure code.
-        chunks         = list(chunk_map.values())[:n_chunks]
-        chunk_ids      = list(chunk_map.keys())[:n_chunks]
-        decoded_chunks = decoder.decode(chunks, chunk_ids)
-
-        # Concatenate the original (decoded) chunks to form the original message
-        decoded_data = b""
-        for chunk in decoded_chunks:
-            assert(len(chunk) == FEC_CHUNK_SIZE)
-            decoded_data += chunk
-
-        # Remove any zero-padding that may have been applied to the last chunk
-        decoded_data = decoded_data[:msg_len]
+        # The encoded data should be decodable at this point
+        if (not decoded_data):
+            raise RuntimeError("Failed to decode the FEC-encoded message")
 
         # We can't know whether the underlying message is encrypted or
         # encapsulated. Thus, for now, put the data into all fields. These
@@ -410,59 +305,15 @@ class ApiMsg:
     def is_fec_decodable(self):
         """Check if the FEC-encoded data is decodable
 
-        Assume that it is decodable when it contains enough FEC chunks.
-
         Returns:
             (bool) Whether the FEC-encoded data is decodable.
 
         """
-        encoded_data = self.data['fec_encoded']
-        assert(encoded_data is not None)
+        assert(self.data['fec_encoded'] is not None)
 
-        # A properly FEC-encoded object contains an integer number of max-sized
-        # payloads. However, this function could be processing any type of data,
-        # not necessarily FEC-encoded. Infer that the object is not decodable if
-        # this formatting condition fails.
-        if (len(encoded_data) % pkt.MAX_PAYLOAD != 0):
-            logger.debug("Not a properly formatted FEC-encoded object")
-            return False
-
-        # Get the original message length and check that it is set consistently
-        # among all payloads. Organize also the set of unique chunk ids within
-        # the internal FEC-encoded data.
-        n_payloads = len(encoded_data) // pkt.MAX_PAYLOAD
-        chunk_ids  = set()
-        msg_len    = None
-        for i_payload in range(n_payloads):
-            # Byte range of the next FEC header
-            s_byte     = i_payload * pkt.MAX_PAYLOAD # starting byte
-            e_byte     = s_byte + FEC_HEADER_LEN # ending byte
-            fec_header = encoded_data[s_byte:e_byte]
-
-            # Unpack the metadata from the FEC header
-            metadata = struct.unpack(FEC_HEADER_FORMAT, fec_header)
-
-            # Add to set of unique chunk ids
-            chunk_ids.add(metadata[0])
-
-            # If the message length is not consistent, assume this object is not
-            # decodable. It could still be another type of data (not
-            # FEC-encoded).
-            if (msg_len is not None and metadata[1] != msg_len):
-                logger.debug("Inconsistent message length on FEC chunks - "
-                             "likely not FEC-encoded")
-                return False
-            else:
-                msg_len = metadata[1]
-
-        # Number of FEC chunks required to decode the original message:
-        n_chunks = ceil(msg_len / FEC_CHUNK_SIZE)
-
-        # Is it decodable?
-        decodable = len(chunk_ids) >= n_chunks
-        logger.debug("FEC object decodable: {}".format(decodable))
-
-        return decodable
+        fec = Fec()
+        res = fec.decode(self.data['fec_encoded'])
+        return res != False
 
     def save(self, dst_dir, target='original'):
         """Save data into a file